fix(docs): move API reference redirect to proxy.ts
Next.js doesn't allow both middleware.ts and proxy.ts.
Added kebab-case → camelCase redirect logic to existing proxy.ts.
